refactor(intake): package intake workflows by feature
Why:
- the flat services directory had accumulated a noisy intake cluster with too many same-prefix modules
- parity-proof tests were also starting to sprawl, which weakens architectural clarity and testability

What:
- move intake scanning, routing, inventory, overlap, package rules, and the main intake service into application/services/intake/
- split intake service and archive scanning into smaller intake-local seams for plan building, report writing, archive models, and archive member handling
- mirror the intake package under tests/unit/application/services/intake/ and split the remaining oversized routing and archive test files
- tighten engineering standards and refresh the master parity ledger to reflect the package-first layout

Checks:
- uv run python -m tools.run_quality_gates --full-tests
