+++ src/crypto_reconciliation/application/services/intake/__init__.py
@@ -1,0 +1,25 @@
+"""Intake workflow services and intake-local helper seams."""
+
+from .archive_scan import ScanIssue, ScannedFile, ScannedTree, scanned_tree_files
+from .file_facts import IntakeFileFacts, detect_capture_id, inspect_intake_file
+from .inventory import resolve_inventory_route
+from .packages import PackageRuleSummary, PlannedPackageItem, apply_package_rules
+from .routing import detect_source_folder, route_intake_file
+from .service import SourceIntakeService
+
+__all__ = [
+    "IntakeFileFacts",
+    "PackageRuleSummary",
+    "PlannedPackageItem",
+    "ScanIssue",
+    "ScannedFile",
+    "ScannedTree",
+    "SourceIntakeService",
+    "apply_package_rules",
+    "detect_capture_id",
+    "detect_source_folder",
+    "inspect_intake_file",
+    "resolve_inventory_route",
+    "route_intake_file",
+    "scanned_tree_files",
+]
