+++ src/crypto_reconciliation/application/services/intake/package_reviews.py
@@ -2,8 +2,8 @@
 
 from __future__ import annotations
 
-from .intake_package_models import BundlePackage, package_key
-from .intake_package_scope import overlap_reason, scope_status
+from .package_models import BundlePackage, package_key
+from .package_scope import overlap_reason, scope_status
 
 
 def apply_default_decisions(
