+++ src/crypto_reconciliation/application/services/manifest.py
@@ -6,7 +6,7 @@ import hashlib
 import json
 
 from crypto_reconciliation.application.dtos import ManifestRequest, ManifestResponse
-from crypto_reconciliation.application.services.archive_scan import scanned_tree_files
+from crypto_reconciliation.application.services.intake import scanned_tree_files
 from crypto_reconciliation.ports.artifacts import ArtifactStorePort
 
 ISSUE_HEADER = ("relative_path", "severity", "kind", "message")
