+++ tests/unit/application/services/intake/test_file_facts.py
@@ -2,12 +2,12 @@ from __future__ import annotations
 
 from pathlib import Path
 
-from crypto_reconciliation.application.services.intake_file_facts import (
+from crypto_reconciliation.application.services.intake.file_facts import (
     IntakeFileFacts,
     detect_capture_id,
     inspect_intake_file,
 )
-from crypto_reconciliation.application.services.intake_routing import detect_source_folder
+from crypto_reconciliation.application.services.intake.routing import detect_source_folder
 
 
 def test_inspect_intake_file_extracts_timestamps_scope_tokens_and_network_hints(tmp_path: Path) -> None:
