+++ tests/unit/application/services/intake/test_package_helpers.py
@@ -3,23 +3,23 @@ from __future__ import annotations
 from collections import Counter
 from datetime import UTC, date, datetime
 
-from crypto_reconciliation.application.services.intake_package_markers import (
+from crypto_reconciliation.application.services.intake.package_markers import (
     extract_datetimes,
     logical_key,
     package_cycle_status,
     package_sort_key,
     row_marker,
 )
-from crypto_reconciliation.application.services.intake_package_models import (
+from crypto_reconciliation.application.services.intake.package_models import (
     BundlePackage,
     PlannedPackageItem,
     package_key,
 )
-from crypto_reconciliation.application.services.intake_package_reviews import (
+from crypto_reconciliation.application.services.intake.package_reviews import (
     apply_default_decisions,
     apply_overlap_review_decisions,
 )
-from crypto_reconciliation.application.services.intake_package_scope import (
+from crypto_reconciliation.application.services.intake.package_scope import (
     compatible_scope,
     material_scope_tokens,
     overlap_reason,
